feat: Move user roles endpoints to Users folder, changes in the routes
