+++ src/Endatix.Api/Endpoints/Users/AssignRoleToUser.AssignRoleRequest.cs
@@ -1,4 +1,4 @@
-namespace Endatix.Api.Endpoints.Auth;
+namespace Endatix.Api.Endpoints.Users;
 
 /// <summary>
 /// Request for assigning a role to a user.
