+++ src/Endatix.Api/Endpoints/Users/AssignRoleToUser.AssignRoleResponse.cs
@@ -1,4 +1,4 @@
-namespace Endatix.Api.Endpoints.Auth;
+namespace Endatix.Api.Endpoints.Users;
 
 /// <summary>
 /// Response for assigning a role to a user.
