+++ src/Endatix.Api/Endpoints/Users/AssignRoleToUser.AssignRoleValidator.cs
@@ -2,7 +2,7 @@ using FastEndpoints;
 using FluentValidation;
 using Endatix.Infrastructure.Identity.Authorization;
 
-namespace Endatix.Api.Endpoints.Auth;
+namespace Endatix.Api.Endpoints.Users;
 
 /// <summary>
 /// Validator for the AssignRoleRequest.
